+++ src/color_group.rs
@@ -1,6 +1,7 @@
 use crate::color::Color;
 use crate::constants::*;
 use crate::cursor::move_cursor_by;
+use crate::OutputFormat;
 use std::io::{self, Write};
 
 pub struct ColorGroup {
@@ -10,7 +11,7 @@ pub struct ColorGroup {
 }
 
 impl ColorGroup {
-    pub fn print_colors(&self, term_width: usize) {
+    fn print_colors_full(&self, term_width: usize) {
         let colors = [&self.tinted, &self.original, &self.shaded];
 
         let mut cursor_x = 0;
@@ -37,5 +38,19 @@ impl ColorGroup {
         }
 
         println!();
+    }
+
+    fn print_colors_basic(&self) {
+        let colors = [&self.original, &self.tinted, &self.shaded];
+        for color in colors {
+            println!("{}: {}, {}", &color.title, color.hex_string(), color.rgb_string());
+        }
+    }
+
+    pub fn print_colors(&self, output_format: OutputFormat, term_width: usize) {
+        match output_format {
+            OutputFormat::Full => self.print_colors_full(term_width),
+            OutputFormat::Basic => self.print_colors_basic()
+        }
     }
 }